refactor: Unkill `ppx_sexp_conv` (#8588)
In https://github.com/returntocorp/semgrep/pull/8196, we removed
`ppx_sexp_conv`. It had an issue
(https://github.com/janestreet/ppx_sexp_conv/issues/40) which made
versions newer than 0.14.3 unusable for us. However, 0.14.3 is
incompatible with OCaml 5.0, so we had to remove it entirely in order to
move towards that.

But, 0.16.0 fixed the issue.

I am working on shipping names/types from the Java standard library with
the Pro Engine, to help with name and type resolution. I need a
shelf-stable serialization format for this -- one where minor changes to
the datatype will either be inconsequential or quickly detected. OCaml's
built-in marshaling is not safe in the face of changes to the datatypes
being marshaled, and can lead to memory corruption issues in these
cases.

In contrast, we will get an error on deserialization if an sexp cannot
be converted back to a valid OCaml value, so I'd like to use
`ppx_sexp_conv` for this.

In this PR, I add `ppx_sexp_conv` version 0.16.0 as a dependency, and
use it on a handful of data structures. The autogenerated code is not
currently used, but I plan to use it in an upcoming change to the Pro
Engine.

Test plan: Automated tests

PR checklist:

- [x] Purpose of the code is [evident to future
readers](https://semgrep.dev/docs/contributing/contributing-code/#explaining-code)
- [x] Tests included or PR comment includes a reproducible test plan
- [x] Documentation is up-to-date
- [x] A changelog entry was [added to
changelog.d](https://semgrep.dev/docs/contributing/contributing-code/#adding-a-changelog-entry)
for any user-facing change
- [x] Change has no security implications (otherwise, ping security
team)

If you're unsure about any of this, please see:

- [Contribution
guidelines](https://semgrep.dev/docs/contributing/contributing-code)!
- [One of the more specific guides located
here](https://semgrep.dev/docs/contributing/contributing/)