feat: match in the presence of wildcard imports (#8514)
## What:
This PR makes it such that patterns that look like qualified names can
take into account "prefixes" due to wildcard imports like `from A import
*` in Python.

## Why:
Previously, matching a qualified name such as `A.B.C.x` was tricky,
because the following programs were all possible:
```
foo(A.B.C.x)
```
```
from A import *
foo(B.C.x)
```
```
from A.B import *
foo(C.x)
```
```
from A.B.C import *
foo(x)
```

However, all of them had different matching behavior, because matching
did not know how to properly take into account the wildcard import,
which makes identifiers within its scope have an implicit prefix to its
qualified name, potentially.

This means that a pattern which matched all four would have to be a
`pattern-either` of all four options. This led to slowdowns when writing
rules, and just general lack of convenience for rule-writers. We want to
make it such that Semgrep understands that the naive pattern `A.B.C.x`
can demonstrate differing behavior, depending on whether there is a
wildcard import around with some prefix of the pattern's qualified name.

## How:
This PR makes it such that the `Matching_generic.env` has a
`wildcard_imports` field in it, which is obtained by visiting the AST of
a target prior to matching, and compiling that list of wildcard imports.

Then, when `m_name` is invoked, a code path has been introduced which
considers the possibility of "unpacking" the wildcard import's prefix.
For instance, if there is a `from A.B import *` somewhere within a
program, we consider unpacking a prefix of `A.B` when matching any name.

This is as described at this scope doc:
https://www.notion.so/Matching-in-the-Presence-of-Wildcard-Imports-fb004e0580874944ac23c084c69be667?d=4a91132287364654934dea529cb2dbab#69cae00fdfc246f2b0a1edcecf86d10a

We chose to put it into the matching environment, because it's quite a
bit of work to consolidate this information at the naming level. To make
this a naming problem, we would have to store the wildcard import
information within our naming context, and update it accordingly to the
AST's scoping rules. Then, we would have to store this information
within the AST itself, which would lead to both probable significant
memory usage, but also an inconvenience in the representation of
identifiers, because this potential prefix might be used at many
identifiers.

Since we expect that scoping of wildcard imports is likely a non-issue
in most practical code, the simplest and cleanest solution is to put it
into the matching environment. It is worth noting that this makes
`Generic_vs_generic` less dependent upon specifically the two nodes
being compared, because there might be "hidden information" floating
around in the environment which alters the match, but I believe this to
be a generally desirable change.

## Test plan:
`make test`

Closes PA-1006

PR checklist:

- [X] Purpose of the code is [evident to future
readers](https://semgrep.dev/docs/contributing/contributing-code/#explaining-code)
- [X] Tests included or PR comment includes a reproducible test plan
- [X] Documentation is up-to-date
- [x] A changelog entry was [added to
changelog.d](https://semgrep.dev/docs/contributing/contributing-code/#adding-a-changelog-entry)
for any user-facing change
- [X] Change has no security implications (otherwise, ping security
team)

If you're unsure about any of this, please see:

- [Contribution
guidelines](https://semgrep.dev/docs/contributing/contributing-code)!
- [One of the more specific guides located
here](https://semgrep.dev/docs/contributing/contributing/)