cleanup: add IL node for fetch expressions used as statements (#8646)
For expression-oriented languages, such as Rust, everything is an
expression. Moreover, the last expression of a function is implicitly
returned.

Currently, the Rust function

```
  fn f(x:i32) {
    x;
    x;
    x;
  }
```

disregards the first two xs, and gives us a CFG with only one return
node for the last x. The return node for the last x is also a hack that
explicitly inserts the node to the last expression statement in a
function.

This change treats expressions as statements in expression-oriented
languages. It generates an IL assigning to a fresh var for these
expression statements.

The tradeoff is creating fresh vars that aren't used, but this gives us
a more complete CFG. This CFG will help when implementing the next
version of implicit return support.

Closes [PA-3056](https://linear.app/semgrep/issue/PA-3056).

Test plan:

Ran `make core-test` and it passes all tests.

Also ran `sc -cfg_il` on the following rust code

```
  fn f(x:i32) {
    x;
    x;
    x;
  }
```

and the CFG shows 2 assignments and one return node. Before this PR, it
only showed the single return node.