fix(secrets): permit both CI and scan config IDs (#8856)
Currently postprocessors are required to pass a validation step in
pysemgrep regarding their origin. This is handled incorrectly in some
cases when we report a config_id in the semgrep ci flow. This resolves
the issue by permitting the config id used for this purpose in addition
to the one associated with registry pro rules.