refactor: kill `ForIn` (#8853)
## What:
This PR kills the `ForIn` construct for `for_header`, preferring instead
the closely related `ForEach`.

## Why:
Both of these constructs have different types, but closely resemble each
other in intention. This was causing us a bug, as a mistranslation to
`ForIn` in C++ caused an inability to do taint tracking properly, which
would not have happened had it been translated to the proper `ForEach`.
See
https://linear.app/semgrep/issue/PA-3090/incorrect-taint-tracking-in-c-range-based-loops

## How:
This PR just modifies the type, and embeds such range-based for loops
into `ForEach` instead. This only takes a little more contorting, which
is particularly noticeable in the case where we need to translate an
entity into a pattern. This was handled relatively broadly, but I didn't
feel like it was worth creating a brand new `pattern_of_entity`
function, as I felt like it was impossible to write truly generally.

## Test plan:
`make test` (notably, the test added in
https://github.com/returntocorp/semgrep/pull/8749 succeeds)

Closes PA-3090 (not just another birthday, it's 30/90...)