Set default timeout to 2s instead of 30s (#8879)
Typically, running a rule on a file should take a fraction of a second.
When a rule takes more than a couple of seconds is often because the
rule is not optimally written *and* the file is unusually large (a
minified file or machine-generated code), so waiting 30s for it does not
tend to bring any value. Plus, by cutting it earlier, we may prevent a
potential OOM crash when running the rule is very memory intensive.
Right now, if we happen to analyze a huge autogenerated file we could
waste up to 90s on it (30s*3). If a well-written rule takes more than 2s
on a normal source file, then we want to know about it and fix it.

Closes PA-3155

Related to PR #8078
See also PR returntocorp/semgrep-rules#3139

test plan:
- Ran `p/default` on 46 repositories from the stress-test-monorepo, the
number of findings is only affected in 6 cases (1-5 findings missing in
each case). The number of timeouts increased in about a dozen cases, in
the majority of them it was 5 or fewer extra timeouts, whereas in a few
cases 6-8 timeouts. Significant speedups on some repos, _brotli_ is ~5x
faster, _juice-shop_ is twice as fast, _langchain_ is ~4x faster, and
_tree-sitter_ is ~10x faster.
- Then try running on a bunch of repos without
`-filter_irrelevant_rules`... This is the case for inter-file analysis,
for example. See also PR #8767.