fix(elixir): make sequential tainting work (#8901)
Before this PR, the statement
```
  x = y
```
is treated as a binary op, and the taint analysis won't properly analyze
the dataflow.

In this PR
```
  x = y
```
is treated as VarDef, and for cases with more involved LHSs, such as
```
  {x1, x2} = {y1, y2}
  [x1, x2] = [y1, y2]
```
are treated as LetPatterns.

In the process, added support to PatList in AST_to_generic, so lists
such as in the example above can be included.

Also added the .ex file type to default to Elixir. This is needed for
looking at the generated IL, and I thought it's useful to keep it.

Tested with `make test-core`.