devops: start to use jsonnet to specify the workflows (#8939)
It's not unusual for devops PRs to need to modify 6 different GHA
workflows
to change a constant. For example
https://github.com/returntocorp/semgrep/pull/8922
to change the OCaml version to use. This is annoying.
We should be able to define a constant file and just
reuse the constant in different workflows.

In the same way YAML (and GHA) forces you sometimes
to duplicate the same thing again and again. See for example
push-docker.yml where the same parameters list need to be
repeated in 2 places just because YAML.

We're programmers, we want to define constants in one place, separate
concerns in different files, reuse code with functions, etc.
So let's use a programming language to specify
our configuration.

Turns out we already know a good programming language
to define configurations: jsonnet!
we decided to switch to jsonnet for Semgrep itself for the exact same
reasons.

And the good news we don't even need to install anything to benefit
from it as Semgrep already contains a jsonnet interpreter (in
tools/ojsonnet).

test plan:
```
$ make build-ojsonnet
$ cd .github/workflows
$ touch semgrep.jsonnet
$ make
=> semgrep.yml regenerated
```